requirements.txt
streamlit
paho-mqtt
opencv-python-headless
pyzbar
qrcode
pillow